# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bowtie API
  version: 0.1.0
  description: Live TV streaming API for HDHomeRun tuners.
paths:
...
  /api/v1/auth/refresh:
    post:
      summary: Rotate a refresh token and issue a new token pair
      operationId: refresh
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access and refresh tokens plus user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/channels:
    get:
      summary: List enabled channels for the viewer
      operationId: listChannels
      tags: [channels]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Enabled channels with optional EPG logo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ViewerChannel'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    User:
      type: object
      required: [id, username, role, maxQuality, maxStreams, maxTuners, allowedChannelIds, maxRating, blockUnrated]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum: [admin, viewer]
        maxQuality:
          type: string
          description: Quality profile name, or empty for unlimited
        maxStreams:
          type: integer
          minimum: 0
          maximum: 8
          description: Concurrent streams allowed (0 = no limit; admins are never limited)
        maxTuners:
          type: integer
          minimum: 0
          maximum: 8
          description: |
            Tuners the account may occupy alone (0 = no limit). Watching a
            channel another account is already watching costs no tuner.
        allowedChannelIds:
          type: array
          nullable: true
          items:
            type: integer
            format: int64
          description: Parental controls — channels this account may see and watch (null = every channel; admins are never restricted)
        maxRating:
          type: string
          enum: ['', TV-Y, TV-Y7, TV-G, TV-PG, TV-14, TV-MA]
          description: Parental controls — highest rating that plays ("" = no limit). Movie ratings map onto the same ladder (G, PG, PG-13, R).
        blockUnrated:
          type: boolean
          description: Parental controls — also block programs without a rating (including channels with no guide data)
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
    ViewerChannel:
      type: object
      required: [id, guideNumber, name, logoUrl, reception, favorite]
      properties:
        favorite:
          type: boolean
          description: The caller has starred this channel
        id:
          type: integer
          format: int64
        guideNumber:
          type: string
        name:
          type: string
        logoUrl:
          type: string
          description: Icon URL from mapped EPG channel, or empty
        reception:
          type: string
          enum: [ok, noSignal, unknown]
          description: |
            The channel's last tune outcome, learned from real tunes (the
            HDHomeRun reports signal only for a channel it is tuned to).
            "noSignal" means the tuner last answered 806/807; "unknown" means
            not tuned since the server started.
        receptionCheckedAt:
          type: string
          format: date-time
          description: When reception was last learned; absent when unknown.
...
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed or missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...